+++ src/extension.ts
@@ -6,7 +6,25 @@ import * as fold from './Fold';
 export function activate(context: vscode.ExtensionContext) {
 
     let disposable = vscode.commands.registerCommand('dakara-foldplus.levelAtCursor', () => {
-        fold.foldUsingLevelOfCursorLine();
+        fold.foldLevelOfCursor();
+    });
+    context.subscriptions.push(disposable);
+
+    disposable = vscode.commands.registerCommand('dakara-foldplus.levelOfParent', () => {
+        fold.foldLevelOfParent();
+    });
+    context.subscriptions.push(disposable);
+
+    disposable = vscode.commands.registerCommand('dakara-foldplus.children', () => {
+        fold.foldChildren();
+    });
+    context.subscriptions.push(disposable);
+
+    disposable = vscode.commands.registerCommand('dakara-foldplus.parent', () => {
+        const textEditor = vscode.window.activeTextEditor;
+        const parentLine = edit.findNextLineUpSpacedLeft(textEditor, textEditor.selection.active.line);
+        textEditor.selection = new vscode.Selection(parentLine.lineNumber, 0, parentLine.lineNumber, 0);
+        vscode.commands.executeCommand('editor.fold');
     });
     context.subscriptions.push(disposable);